[file-service] throw CannotDeleteFile if doesn\'t exists

+++ src/Http/Services/FileService.php
@@ -90,10 +90,14 @@ class FileService implements FileServiceContract
             $path = $url;
         }
         try {
-            if ($this->disk->mimeType($path) === 'directory') {
-                $deleted = $this->disk->deleteDirectory($path);
+            if ($this->disk->exists($path)) {
+                if ($this->disk->mimeType($path) === 'directory') {
+                    $deleted = $this->disk->deleteDirectory($path);
+                } else {
+                    $deleted = $this->disk->delete($path);
+                }
             } else {
-                $deleted = $this->disk->delete($path);
+                $deleted = false;
             }
             if (!$deleted) {
                 throw new CannotDeleteFile($url);
